+++ project/app/src/main/java/com/chalmers/tda367/localfeud/net/RestClient.java
@@ -1,8 +1,8 @@
-package com.chalmers.tda367.localfeud.util;
+package com.chalmers.tda367.localfeud.net;
 
 import android.util.Log;
 
-import com.chalmers.tda367.localfeud.net.IResponseAction;
+import com.chalmers.tda367.localfeud.util.TagHandler;
 import com.loopj.android.http.AsyncHttpClient;
 import com.loopj.android.http.RequestParams;
