+++ project/app/src/main/java/com/chalmers/tda367/localfeud/net/RestResponseHandler.java
@@ -1,8 +1,9 @@
-package com.chalmers.tda367.localfeud.util;
+package com.chalmers.tda367.localfeud.net;
 
 import android.util.Log;
 
 import com.chalmers.tda367.localfeud.net.IResponseAction;
+import com.chalmers.tda367.localfeud.util.TagHandler;
 import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.loopj.android.http.ResponseHandlerInterface;
